Refactor to layout components folder
Move Layout components to components/Layout
