+++ src/components/Blocks/BlocksList.tsx
@@ -2,13 +2,13 @@ import { generatePath, useNavigate, useParams } from 'react-router-dom'
 import { RoutedPagination } from '~components/Pagination/Pagination'
 import { RoutedPaginationProvider } from '~components/Pagination/PaginationProvider'
 import { PaginationItemsPerPage, RoutePath } from '~constants'
-import { LoadingCards } from '~src/layout/Loading'
-import LoadingError from '~src/layout/LoadingError'
+import { LoadingCards } from '~components/Layout/Loading'
+import LoadingError from '~components/Layout/LoadingError'
 import { useBlockList } from '~queries/blocks'
 import { useChainInfo } from '~queries/stats'
 import { BlockCard } from '~components/Blocks/BlockCard'
 import { useTranslation } from 'react-i18next'
-import { InputSearch } from '~src/layout/Inputs'
+import { InputSearch } from '~components/Layout/Inputs'
 
 export const BlocksFilter = () => {
   const { t } = useTranslation()
