+++ src/components/Layout/ShowRawButton.tsx
@@ -1,7 +1,7 @@
 import { Box, BoxProps, Button, ButtonProps, useDisclosure } from '@chakra-ui/react'
 import { Trans } from 'react-i18next'
 import { CopyButtonIcon } from '~components/CopyButton'
-import { JsonViewer } from './JsonViewer'
+import { JsonViewer } from '~components/Layout/JsonViewer'
 
 const ShowRawButton = ({ obj, ...props }: { obj: object } & Omit<ButtonProps, 'onClick'>) => {
   const { getDisclosureProps, getButtonProps } = useDisclosure()
