+++ src/components/Process/Detail.tsx
@@ -34,15 +34,15 @@ import {
   PublishedElection,
 } from '@vocdoni/sdk'
 import { FallbackHeaderImg, RoutePath } from '~constants'
-import { HeroHeaderLayout } from '~src/layout/HeroHeaderLayout'
+import { HeroHeaderLayout } from '~components/Layout/HeroHeaderLayout'
 import { CopyButton, ReducedTextAndCopy } from '~components/CopyButton'
 import { Trans, useTranslation } from 'react-i18next'
 import { ElectionStatusBadge } from '~components/Organizations/StatusBadge'
 import { OrganizationCard } from '~components/Organizations/Card'
-import { RawContentBox } from '~src/layout/ShowRawButton'
+import { RawContentBox } from '~components/Layout/ShowRawButton'
 import { useElectionKeys, useElectionVotesList } from '~queries/processes'
 import { PaginationProvider, usePagination } from '~components/Pagination/PaginationProvider'
-import { LoadingCards } from '~src/layout/Loading'
+import { LoadingCards } from '~components/Layout/Loading'
 import { Pagination } from '~components/Pagination/Pagination'
 import { BiEnvelope } from 'react-icons/bi'
 import { ucfirst } from '~utils/strings'
