+++ src/components/Process/ProcessList.tsx
@@ -6,9 +6,9 @@ import { RoutedPagination } from '~components/Pagination/Pagination'
 import { RoutedPaginationProvider } from '~components/Pagination/PaginationProvider'
 import { PaginationItemsPerPage, RoutePath } from '~constants'
 import { useProcessesCount, useProcessList } from '~queries/processes'
-import { InputSearch } from '~src/layout/Inputs'
-import { LoadingCards } from '~src/layout/Loading'
-import LoadingError from '~src/layout/LoadingError'
+import { InputSearch } from '~components/Layout/Inputs'
+import { LoadingCards } from '~components/Layout/Loading'
+import LoadingError from '~components/Layout/LoadingError'
 import useQueryParams from '~src/router/use-query-params'
 import { isEmpty } from '~utils/objects'
 import { ElectionCard } from './Card'
