+++ src/components/Verify/index.tsx
@@ -3,7 +3,7 @@ import { useState } from 'react'
 import { Trans, useTranslation } from 'react-i18next'
 import { generatePath, useNavigate, useParams } from 'react-router-dom'
 import { RoutePath } from '~constants'
-import { InputSearch } from '~src/layout/Inputs'
+import { InputSearch } from '~components/Layout/Inputs'
 import addVote from '/images/add-vote.svg'
 
 const SearchVote = ({ compact }: { compact?: boolean }) => {
