+++ src/pages/blocks.tsx
@@ -1,4 +1,4 @@
-import ListPageLayout from '~src/layout/ListPageLayout'
+import ListPageLayout from '~components/Layout/ListPageLayout'
 import { useTranslation } from 'react-i18next'
 import { useChainInfo } from '~queries/stats'
 import { BlocksFilter, PaginatedBlocksList } from '~components/Blocks/BlocksList'
