+++ src/pages/organizations.tsx
@@ -1,5 +1,5 @@
 import { OrganizationsFilter, PaginatedOrganizationsList } from '~components/Organizations/List'
-import ListPageLayout from '~src/layout/ListPageLayout'
+import ListPageLayout from '~components/Layout/ListPageLayout'
 import { useOrganizationCount } from '~queries/organizations'
 import { useTranslation } from 'react-i18next'
