+++ src/router/SuspenseLoader.tsx
@@ -1,5 +1,5 @@
 import { ReactNode, Suspense } from 'react'
-import { Loading } from '~src/layout/Loading'
+import { Loading } from '~components/Layout/Loading'
 
 export const SuspenseLoader = ({ children }: { children: ReactNode }) => (
   <Suspense fallback={<Loading />}>{children}</Suspense>
